+++ packages/components-js/tests/e2e/specs/radio-button-wrapper.e2e.ts
@@ -1,99 +1,119 @@
 import {
+  getActiveElementId,
+  getActiveElementTagName,
   getAttribute,
-  getCssClasses, getElementStyle, getProperty,
+  getBrowser,
+  getCssClasses,
+  getElementStyle,
+  getProperty,
   selectNode,
-  setContentWithDesignSystem, waitForStencilLifecycle
+  setContentWithDesignSystem,
+  waitForStencilLifecycle
 } from '../helpers';
 import { Page } from 'puppeteer';
-import { getBrowser } from '../helpers/setup';
 
 describe('radio-button-wrapper', () => {
-
   let page: Page;
 
-  beforeEach(async () => page = await getBrowser().newPage());
+  beforeEach(async () => (page = await getBrowser().newPage()));
   afterEach(async () => await page.close());
 
   const getRadioButtonHost = () => selectNode(page, 'p-radio-button-wrapper');
-  const getRadioButtonFakeInput = () => selectNode(page, 'p-radio-button-wrapper >>> .p-radio-button-wrapper__fake-radio-button');
+  const getRadioButtonFakeInput = () =>
+    selectNode(page, 'p-radio-button-wrapper >>> .p-radio-button-wrapper__fake-radio-button');
   const getRadioButtonRealInput = () => selectNode(page, 'p-radio-button-wrapper input');
   const getRadioButtonLabel = () => selectNode(page, 'p-radio-button-wrapper >>> .p-radio-button-wrapper__label-text');
   const getRadioButtonMessage = () => selectNode(page, 'p-radio-button-wrapper >>> .p-radio-button-wrapper__message');
 
   it('should render', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label">
         <input type="radio" name="some-name"/>
       </p-radio-button-wrapper>
-    `);
+    `
+    );
     const el = await getRadioButtonFakeInput();
     expect(el).toBeDefined();
   });
 
   it('should add aria-label to support screen readers properly', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label">
         <input type="radio" name="some-name"/>
       </p-radio-button-wrapper>
-    `);
+    `
+    );
     const input = await getRadioButtonRealInput();
     expect(await getProperty(input, 'ariaLabel')).toBe('Some label');
   });
 
   it('should add aria-label with message text to support screen readers properly', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label" message="Some error message" state="error">
         <input type="radio" name="some-name"/>
       </p-radio-button-wrapper>
-    `);
+    `
+    );
     const input = await getRadioButtonRealInput();
     expect(await getProperty(input, 'ariaLabel')).toBe('Some label. Some error message');
   });
 
   it('should not render label if label prop is not defined but should render if changed programmatically', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper>
         <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+      </p-radio-button-wrapper>`
+    );
 
     const radioComponent = await getRadioButtonHost();
     expect(await getRadioButtonLabel()).toBeNull();
 
-    await radioComponent.evaluate(el => el.setAttribute('label', 'Some label'));
+    await radioComponent.evaluate((el) => el.setAttribute('label', 'Some label'));
     await waitForStencilLifecycle(page);
 
     expect(await getRadioButtonLabel()).not.toBeNull();
   });
 
   it('should add/remove message text and update aria-label attribute with message if state changes programmatically', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label">
         <input type="radio" name="some-name"/>
         <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+      </p-radio-button-wrapper>`
+    );
 
     const radioComponent = await getRadioButtonHost();
     const input = await selectNode(page, 'input');
 
     expect(await getRadioButtonMessage()).toBeNull();
-    await radioComponent.evaluate(el => el.setAttribute('state', 'error'));
-    await radioComponent.evaluate(el => el.setAttribute('message', 'Some error message'));
+    await radioComponent.evaluate((el) => el.setAttribute('state', 'error'));
+    await radioComponent.evaluate((el) => el.setAttribute('message', 'Some error message'));
     await waitForStencilLifecycle(page);
 
     expect(await getRadioButtonMessage()).toBeDefined();
     expect(await getAttribute(await getRadioButtonMessage(), 'role')).toEqual('alert');
     expect(await getProperty(input, 'ariaLabel')).toEqual('Some label. Some error message');
 
-    await radioComponent.evaluate(el => el.setAttribute('state', 'success'));
-    await radioComponent.evaluate(el => el.setAttribute('message', 'Some success message'));
+    await radioComponent.evaluate((el) => el.setAttribute('state', 'success'));
+    await radioComponent.evaluate((el) => el.setAttribute('message', 'Some success message'));
     await waitForStencilLifecycle(page);
 
     expect(await getRadioButtonMessage()).toBeDefined();
     expect(await getAttribute(await getRadioButtonMessage(), 'role')).toBeNull();
     expect(await getProperty(input, 'ariaLabel')).toEqual('Some label. Some success message');
 
-    await radioComponent.evaluate(el => el.setAttribute('state', 'none'));
-    await radioComponent.evaluate(el => el.setAttribute('message', ''));
+    await radioComponent.evaluate((el) => el.setAttribute('state', 'none'));
+    await radioComponent.evaluate((el) => el.setAttribute('message', ''));
     await waitForStencilLifecycle(page);
 
     expect(await getRadioButtonMessage()).toBeNull();
@@ -101,13 +121,16 @@ describe('radio-button-wrapper', () => {
   });
 
   it('should check radio-button when input is clicked', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label" id="radio-1">
         <input type="radio" name="some-name"/>
       </p-radio-button-wrapper>
       <p-radio-button-wrapper label="Some label" id="radio-2">
         <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+      </p-radio-button-wrapper>`
+    );
 
     const radioComponent1Label = await selectNode(page, '#radio-1 >>> .p-radio-button-wrapper__label');
     const fakeRadio1 = await selectNode(page, '#radio-1 >>> .p-radio-button-wrapper__fake-radio-button');
@@ -130,13 +153,16 @@ describe('radio-button-wrapper', () => {
   });
 
   it('should check radio-button when label text is clicked', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label" id="radio-1">
-        <input type="radio" name="some-name"/>
+        <input id="radio-1-input" type="radio" name="some-name"/>
       </p-radio-button-wrapper>
       <p-radio-button-wrapper label="Some label" id="radio-2">
-        <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+        <input id="radio-2-input" type="radio" name="some-name"/>
+      </p-radio-button-wrapper>`
+    );
 
     const radioComponent1Label = await selectNode(page, '#radio-1 >>> .p-radio-button-wrapper__label');
     const fakeRadio1 = await selectNode(page, '#radio-1 >>> .p-radio-button-wrapper__fake-radio-button');
@@ -145,27 +171,34 @@ describe('radio-button-wrapper', () => {
     const labelText2 = await selectNode(page, '#radio-2 >>> .p-radio-button-wrapper__label-text');
 
     expect(await getCssClasses(fakeRadio1)).not.toContain('p-radio-button-wrapper__fake-radio-button--checked');
+    expect(await getActiveElementId(page)).toBe('');
+    expect(await getActiveElementTagName(page)).toBe('BODY');
 
     await labelText1.click();
     await waitForStencilLifecycle(page);
 
     expect(await getCssClasses(fakeRadio1)).toContain('p-radio-button-wrapper__fake-radio-button--checked');
+    expect(await getActiveElementId(page)).toBe('radio-1-input');
 
     await labelText2.click();
     await waitForStencilLifecycle(page);
 
     expect(await getCssClasses(fakeRadio1)).not.toContain('p-radio-button-wrapper__fake-radio-button--checked');
     expect(await getCssClasses(fakeRadio2)).toContain('p-radio-button-wrapper__fake-radio-button--checked');
+    expect(await getActiveElementId(page)).toBe('radio-2-input');
   });
 
   it('should check radio-button when radio-button is changed programmatically', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label" id="radio-1">
         <input type="radio" name="some-name"/>
       </p-radio-button-wrapper>
       <p-radio-button-wrapper label="Some label" id="radio-2">
         <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+      </p-radio-button-wrapper>`
+    );
 
     const fakeRadio1 = await selectNode(page, '#radio-1 >>> .p-radio-button-wrapper__fake-radio-button');
     const fakeRadio1Input = await selectNode(page, '#radio-1 > input');
@@ -175,12 +208,12 @@ describe('radio-button-wrapper', () => {
     expect(await getCssClasses(fakeRadio1)).not.toContain('p-radio-button-wrapper__fake-radio-button--checked');
     expect(await getCssClasses(fakeRadio2)).not.toContain('p-radio-button-wrapper__fake-radio-button--checked');
 
-    await fakeRadio1Input.evaluate(el => el.setAttribute('checked', 'true'));
+    await fakeRadio1Input.evaluate((el) => el.setAttribute('checked', 'true'));
     await waitForStencilLifecycle(page);
 
     expect(await getCssClasses(fakeRadio1)).toContain('p-radio-button-wrapper__fake-radio-button--checked');
 
-    await fakeRadio2Input.evaluate(el => el.setAttribute('checked', 'true'));
+    await fakeRadio2Input.evaluate((el) => el.setAttribute('checked', 'true'));
     await waitForStencilLifecycle(page);
 
     expect(await getCssClasses(fakeRadio1)).not.toContain('p-radio-button-wrapper__fake-radio-button--checked');
@@ -188,48 +221,58 @@ describe('radio-button-wrapper', () => {
   });
 
   it('should disable radio-button when radio-button is set disabled programmatically', async () => {
-    await setContentWithDesignSystem(page, `
+    await setContentWithDesignSystem(
+      page,
+      `
       <p-radio-button-wrapper label="Some label" id="radio-1">
         <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+      </p-radio-button-wrapper>`
+    );
 
     const fakeRadio1 = await selectNode(page, '#radio-1 >>> .p-radio-button-wrapper__fake-radio-button');
     const fakeRadio1Input = await selectNode(page, '#radio-1 > input');
 
     expect(await getCssClasses(fakeRadio1)).not.toContain('p-radio-button-wrapper__fake-radio-button--disabled');
 
-    await fakeRadio1Input.evaluate((el: HTMLInputElement) => el.disabled = true);
+    await fakeRadio1Input.evaluate((el: HTMLInputElement) => (el.disabled = true));
     await waitForStencilLifecycle(page);
 
     expect(await getCssClasses(fakeRadio1)).toContain('p-radio-button-wrapper__fake-radio-button--disabled');
 
-    await fakeRadio1Input.evaluate((el: HTMLInputElement) => el.disabled = false);
+    await fakeRadio1Input.evaluate((el: HTMLInputElement) => (el.disabled = false));
     await waitForStencilLifecycle(page);
 
     expect(await getCssClasses(fakeRadio1)).not.toContain('p-radio-button-wrapper__fake-radio-button--disabled');
   });
 
   describe('hover state', () => {
-
     it('should change box-shadow color when fake radio button is hovered', async () => {
-      await setContentWithDesignSystem(page, `
+      await setContentWithDesignSystem(
+        page,
+        `
       <p-radio-button-wrapper label="Some label" id="radio-1">
         <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+      </p-radio-button-wrapper>`
+      );
 
       const fakeRadioButton = await getRadioButtonFakeInput();
       const initialBoxShadow = await getElementStyle(fakeRadioButton, 'boxShadow');
 
       await fakeRadioButton.hover();
 
-      expect(await getElementStyle(fakeRadioButton, 'boxShadow', { waitForTransition: true })).not.toBe(initialBoxShadow);
+      expect(await getElementStyle(fakeRadioButton, 'boxShadow', { waitForTransition: true })).not.toBe(
+        initialBoxShadow
+      );
     });
 
     it('should change box-shadow color of fake radio button when label text is hovered', async () => {
-      await setContentWithDesignSystem(page, `
+      await setContentWithDesignSystem(
+        page,
+        `
       <p-radio-button-wrapper label="Some label" id="radio-1">
         <input type="radio" name="some-name"/>
-      </p-radio-button-wrapper>`);
+      </p-radio-button-wrapper>`
+      );
 
       const fakeRadioButton = await getRadioButtonFakeInput();
       const labelText = await getRadioButtonLabel();
@@ -238,7 +281,9 @@ describe('radio-button-wrapper', () => {
 
       await labelText.hover();
 
-      expect(await getElementStyle(fakeRadioButton, 'boxShadow', { waitForTransition: true })).not.toBe(initialBoxShadow);
+      expect(await getElementStyle(fakeRadioButton, 'boxShadow', { waitForTransition: true })).not.toBe(
+        initialBoxShadow
+      );
     });
   });
 });