+++ packages/storefront/tests/unit/components.code-block.spec.ts
@@ -1,26 +1,57 @@
-import { shallowMount } from '@vue/test-utils';
+import { shallowMount, Wrapper } from '@vue/test-utils';
 import CodeBlock from '@/components/CodeBlock.vue';
+import Vuex, { Store } from 'vuex';
+import Vue from 'vue';
+import { State } from '@/store';
+import { Framework } from '@/models';
 
-async function tick(timeout = 100) {
-  await new Promise(resolve => {
+const tick = async (timeout = 100) => {
+  await new Promise((resolve) => {
     setTimeout(resolve, timeout);
   });
-}
+};
+
+type PartialState = Pick<State, 'selectedFramework'>;
 
 describe('CodeBlock.vue', () => {
+  Vue.use(Vuex);
+
+  const stubs = ['p-text'];
+  const store: Store<PartialState> = new Vuex.Store({
+    state: {
+      selectedFramework: 'vanilla-js'
+    },
+    getters: {
+      selectedFramework: (state: PartialState) => state.selectedFramework
+    },
+    mutations: {
+      setSelectedFramework: (state: PartialState, payload: Framework): void => {
+        state.selectedFramework = payload;
+      }
+    }
+  });
+
+  const getTabButton = (wrapper: Wrapper<Vue>, type: 'vanillajs' | 'angular' | 'react'): Wrapper<Vue> =>
+    wrapper.find(`.tabs .tab:nth-child(${['vanillajs', 'angular', 'react'].indexOf(type) + 1}) button`);
+
+  beforeEach(() => {
+    store.commit('setSelectedFramework', 'vanilla-js');
+  });
+
   it('should render Vanilla JS code block by default', () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag some-attribute="some value">
+        markup: `<p-some-tag some-attribute="some value">
   <span>some text</span>
 </p-some-tag>`
       }
     });
 
-    expect(wrapper.find('.tabs .tab:nth-child(1) button').text()).toBe('Vanilla JS');
-    expect(wrapper.find('.tabs .tab:nth-child(1) button').classes()).toContain('is-active');
+    const btn = getTabButton(wrapper, 'vanillajs');
+    expect(btn.text()).toBe('Vanilla JS');
+    expect(btn.classes()).toContain('is-active');
     expect(wrapper.find('code').text()).toBe(
       `<p-some-tag some-attribute="some value">
   <span>some text</span>
@@ -30,23 +61,24 @@ describe('CodeBlock.vue', () => {
 
   it('should render Angular code block after selecting it', async () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag some-attribute="some value">
+        markup: `<p-some-tag some-attribute="some value">
   <span>some text</span>
 </p-some-tag>`
       }
     });
 
-    expect(wrapper.find('.tabs .tab:nth-child(2) button').text()).toBe('Angular');
-    expect(wrapper.find('.tabs .tab:nth-child(2) button').classes()).not.toContain('is-active');
+    const btn = getTabButton(wrapper, 'angular');
+    expect(btn.text()).toBe('Angular');
+    expect(btn.classes()).not.toContain('is-active');
 
-    wrapper.find('.tabs .tab:nth-child(2) button').trigger('click');
+    btn.trigger('click');
 
     await tick();
 
-    expect(wrapper.find('.tabs .tab:nth-child(2) button').classes()).toContain('is-active');
+    expect(btn.classes()).toContain('is-active');
     expect(wrapper.find('code').text()).toBe(
       `<p-some-tag [someAttribute]="'some value'">
   <span>some text</span>
@@ -56,51 +88,55 @@ describe('CodeBlock.vue', () => {
 
   it('should render React code block after selecting it', async () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag some-attribute="some value">
+        markup: `<p-some-tag some-attribute="some value">
   <span>some text</span>
 </p-some-tag>`
       }
     });
 
-    expect(wrapper.find('.tabs .tab:nth-child(3) button').text()).toBe('React');
-    expect(wrapper.find('.tabs .tab:nth-child(3) button').classes()).not.toContain('is-active');
+    const btn = getTabButton(wrapper, 'react');
+    expect(btn.text()).toBe('React');
+    expect(btn.classes()).not.toContain('is-active');
 
-    wrapper.find('.tabs .tab:nth-child(3) button').trigger('click');
+    btn.trigger('click');
 
     await tick();
 
-    expect(wrapper.find('.tabs .tab:nth-child(3) button').classes()).toContain('is-active');
+    expect(btn.classes()).toContain('is-active');
     expect(wrapper.find('code').text()).toBe(
-      `<PSomeTag someAttribute={"some value"}>
+      `<PSomeTag someAttribute="some value">
   <span>some text</span>
-</PSomeTag>`);
+</PSomeTag>`
+    );
   });
 
   it('should theme code block', async () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text']
+      stubs,
+      store
     });
 
-    expect(wrapper.find('.code-block').classes()).not.toContain('dark');
-    expect(wrapper.find('.code-block').classes()).toContain('light');
+    const block = wrapper.find('.code-block');
+    expect(block.classes()).not.toContain('dark');
+    expect(block.classes()).toContain('light');
 
-    wrapper.setProps({theme: 'dark'});
+    wrapper.setProps({ theme: 'dark' });
 
     await tick();
 
-    expect(wrapper.find('.code-block').classes()).not.toContain('light');
-    expect(wrapper.find('.code-block').classes()).toContain('dark');
+    expect(block.classes()).not.toContain('light');
+    expect(block.classes()).toContain('dark');
   });
 
   it('should remove empty comments', () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag some-attribute="some value" class="some-class">
+        markup: `<p-some-tag some-attribute="some value" class="some-class">
   <!----><span>some text</span><!-- some comment -->
 </p-some-tag>`
       }
@@ -115,10 +151,10 @@ describe('CodeBlock.vue', () => {
 
   it('should beautify code block', () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag     some-attribute="some value"    class="some-class"   >
+        markup: `<p-some-tag     some-attribute="some value"    class="some-class"   >
                           <span       >some text</span>
                 </p-some-tag>`
       }
@@ -133,10 +169,10 @@ describe('CodeBlock.vue', () => {
 
   it('should remove Vue JS specific attributes from code block', () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag data-v-c6a10ac0="" data-v-8dbc1b2a='' data-v-7ba5bd90 some-attribute="some value" class="some-class">
+        markup: `<p-some-tag data-v-c6a10ac0="" data-v-8dbc1b2a='' data-v-7ba5bd90 some-attribute="some value" class="some-class">
   <span>some text</span>
 </p-some-tag>`
       }
@@ -151,10 +187,10 @@ describe('CodeBlock.vue', () => {
 
   it('should remove default web components attributes from code block', () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag theme="light" some-attribute="some value" class="some-class"></p-some-tag>
+        markup: `<p-some-tag theme="light" some-attribute="some value" class="some-class"></p-some-tag>
 <p-some-tag theme="light"></p-some-tag>
 <p-some-tag theme="dark"></p-some-tag>`
       }
@@ -169,10 +205,10 @@ describe('CodeBlock.vue', () => {
 
   it('should remove Stencil JS css classes from code block', async () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag some-attribute="some value" class="hydrated">
+        markup: `<p-some-tag some-attribute="some value" class="hydrated">
   <span>some text</span>
 </p-some-tag>`
       }
@@ -185,8 +221,7 @@ describe('CodeBlock.vue', () => {
     );
 
     wrapper.setProps({
-      markup:
-        `<p-some-tag some-attribute="some value" class="some-class hydrated another-class">
+      markup: `<p-some-tag some-attribute="some value" class="some-class hydrated another-class">
   <span>some text</span>
 </p-some-tag>`
     });
@@ -202,17 +237,17 @@ describe('CodeBlock.vue', () => {
 
   it('should convert code block to Angular syntax', async () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag some-attribute="some value" attribute="some value" class="some-class" another-attribute="{ bar: 'foo' }" onclick="alert('click'); return false;" digit-attribute="6" boolean-attribute="true">
+        markup: `<p-some-tag some-attribute="some value" attribute="some value" class="some-class" another-attribute="{ bar: 'foo' }" onclick="alert('click'); return false;" digit-attribute="6" boolean-attribute="true">
   <span>some text</span>
   <span slot="some-slot">some slot text</span>
 </p-some-tag>`
       }
     });
 
-    wrapper.find('.tabs .tab:nth-child(2) button').trigger('click');
+    getTabButton(wrapper, 'angular').trigger('click');
 
     await tick();
 
@@ -226,23 +261,104 @@ describe('CodeBlock.vue', () => {
 
   it('should convert code block to React syntax', async () => {
     const wrapper = shallowMount(CodeBlock, {
-      stubs: ['p-text'],
+      stubs,
+      store,
       propsData: {
-        markup:
-          `<p-some-tag some-attribute="some value" attribute="some value" class="some-class" another-attribute="{ bar: 'foo' }" onclick="alert('click'); return false;" digit-attribute="6" boolean-attribute="true">
+        markup: `<p-some-tag some-attribute="some value" attribute="some value" class="some-class" another-attribute="{ bar: 'foo' }" onclick="alert('click'); return false;" digit-attribute="6" boolean-attribute="true">
   <span>some text</span>
 </p-some-tag>`
       }
     });
 
-    wrapper.find('.tabs .tab:nth-child(3) button').trigger('click');
+    getTabButton(wrapper, 'react').trigger('click');
 
     await tick();
 
     expect(wrapper.find('code').text()).toBe(
-      `<PSomeTag someAttribute={"some value"} attribute={"some value"} className={"some-class"} anotherAttribute={{ bar: 'foo' }} onClick={()=> {alert('click'); return false;}} digitAttribute={6} booleanAttribute={true}>
+      `<PSomeTag someAttribute="some value" attribute="some value" className="some-class" anotherAttribute={{ bar: 'foo' }} onClick={()=> {alert('click'); return false;}} digitAttribute={6} booleanAttribute={true}>
   <span>some text</span>
 </PSomeTag>`
     );
+  });
+
+  it('should add closing slash on input tags', () => {
+    const wrapper = shallowMount(CodeBlock, {
+      stubs,
+      store,
+      propsData: {
+        markup: `<input type="checkbox">`
+      }
+    });
+
+    expect(wrapper.find('code').text()).toBe(`<input type="checkbox" />`);
+  });
+
+  it('should add new line between tags for readability', () => {
+    const wrapper = shallowMount(CodeBlock, {
+      stubs,
+      store,
+      propsData: {
+        markup: `<div><span></span></div>`
+      }
+    });
+
+    expect(wrapper.find('code').text()).toBe(`<div>
+  <span></span>
+</div>`);
+  });
+
+  it('should remove new line when tag closes immediately', () => {
+    const wrapper = shallowMount(CodeBlock, {
+      stubs,
+      store,
+      propsData: {
+        markup: `<div>
+</div>`
+      }
+    });
+
+    expect(wrapper.find('code').text()).toBe(`<div></div>`);
+  });
+
+  it('should remove br tags', () => {
+    const wrapper = shallowMount(CodeBlock, {
+      stubs,
+      store,
+      propsData: {
+        markup: `<div></div><br><div></div><br><div></div>`
+      }
+    });
+
+    expect(wrapper.find('code').text()).toBe(`<div></div>
+<div></div>
+<div></div>`);
+  });
+
+  it('should replace multiple new lines', () => {
+    const wrapper = shallowMount(CodeBlock, {
+      stubs,
+      store,
+      propsData: {
+        markup: `<div></div><br><br><div></div><br><br><br><div></div>`
+      }
+    });
+
+    expect(wrapper.find('code').text()).toBe(`<div></div>
+
+<div></div>
+
+<div></div>`);
+  });
+
+  it('should clean checked, disabled and readonly attributes', () => {
+    const wrapper = shallowMount(CodeBlock, {
+      stubs,
+      store,
+      propsData: {
+        markup: `<input type="checkbox" checked="checked" disabled="disabled" readonly="readonly">`
+      }
+    });
+
+    expect(wrapper.find('code').text()).toBe(`<input type="checkbox" checked disabled readonly />`);
   });
 });